% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_functions.R
\name{add_plot_coordinates}
\alias{add_plot_coordinates}
\title{Add 1ha IRd plot coordinates}
\usage{
add_plot_coordinates(
  dataset,
  ddlat = "Latitude",
  ddlon = "Longitude",
  launch_add_data = FALSE,
  X_theo = "X_theo",
  Y_theo = "Y_theo",
  check_existing_data = TRUE,
  add_cols = NULL,
  cor_cols = NULL,
  collector_field = NULL
)
}
\arguments{
\item{dataset}{tibble}

\item{ddlat}{column name of dataset containing latitude in decimal degrees}

\item{ddlon}{column name of dataset containing longitude in decimal degrees}

\item{launch_add_data}{whether addd data or not}

\item{X_theo}{column that contain the X quadrat name}

\item{Y_theo}{column that contain the Y quadrat name}

\item{check_existing_data}{check if data already exists}

\item{add_cols}{string character vectors with columns names of dataset of additonal information}

\item{cor_cols}{string character vectors with colums names corresponding to add_cols}

\item{collector_field}{string vector of size one with column name containing the name of the person collecting data}
}
\value{
print html in viewer
}
\description{
print table as html in viewer reordered
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
