% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_functions.R
\name{add_specimens}
\alias{add_specimens}
\title{Add new specimens data}
\usage{
add_specimens(
  new_data,
  col_names_select,
  col_names_corresp,
  plot_name_field = NULL,
  collector_field = NULL,
  launch_adding_data = FALSE
)
}
\arguments{
\item{new_data}{tibble new data to be imported}

\item{col_names_select}{string plot name of the selected plots}

\item{col_names_corresp}{string country of the selected plots}

\item{plot_name_field}{integer indicate which name of col_names_select is the id for matching liste plots table}

\item{collector_field}{integer indicate which name of col_names_select is the id for matching collector}

\item{launch_adding_data}{logical FALSE whether adding should be done or not}
}
\value{
No return value individuals updated
}
\description{
Add new specimens data
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
