% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_manip_db.R
\name{add_taxa_table_taxa}
\alias{add_taxa_table_taxa}
\title{List, extract taxa}
\usage{
add_taxa_table_taxa(ids = NULL)
}
\arguments{
\item{ids}{idtax to filter the table}
}
\value{
A tibble of all taxa

A src object with merged taxa
}
\description{
Provide list of selected taxa
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
