% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_functions.R
\name{add_traits_measures}
\alias{add_traits_measures}
\title{Add an observation in trait measurement table}
\usage{
add_traits_measures(
  new_data,
  col_names_select = NULL,
  col_names_corresp = NULL,
  collector_field = NULL,
  plot_name_field = NULL,
  individual_plot_field = NULL,
  id_plot_name = NULL,
  id_tag_plot = NULL,
  id_specimen = NULL,
  traits_field,
  features_field = NULL,
  allow_multiple_value = FALSE,
  add_data = FALSE
)
}
\arguments{
\item{new_data}{tibble}

\item{col_names_select}{string vector}

\item{col_names_corresp}{string vector}

\item{collector_field}{string column name which contain the collector name}

\item{plot_name_field}{string column name which contain the plot_name for linking}

\item{individual_plot_field}{string column name which contain the individual tag for linking}

\item{id_plot_name}{string column name which contain the ID of plot_name}

\item{id_tag_plot}{string column name which contain the ID of individuals table}

\item{id_specimen}{string column name which contain the ID of specimen}

\item{traits_field}{string vector listing trait columns names in new_data}

\item{features_field}{string vector listing features (column names) to link to measurementsin new_data}

\item{allow_multiple_value}{if multiple values linked to one individual can be uploaded at once}

\item{add_data}{logical whether or not data should be added - by default FALSE}
}
\value{
list of tibbles that should be/have been added
}
\description{
Add a trait measure in trait measurement table
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
