% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsplots_features_function.R
\name{aggregate_character_plot_features}
\alias{aggregate_character_plot_features}
\title{Aggregate character plot features}
\usage{
aggregate_character_plot_features(data)
}
\description{
Aggregate character plot features
}
\keyword{internal}
