% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsplots_features_function.R
\name{aggregate_numeric_plot_features}
\alias{aggregate_numeric_plot_features}
\title{Aggregate numeric plot features}
\usage{
aggregate_numeric_plot_features(data)
}
\description{
Aggregate numeric plot features
}
\keyword{internal}
