% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsplots_features_function.R
\name{aggregate_plot_features}
\alias{aggregate_plot_features}
\title{Aggregate plot features to wide format using data.table}
\usage{
aggregate_plot_features(data, con)
}
\description{
Aggregate plot features to wide format using data.table
}
\keyword{internal}
