% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsplots_features_function.R
\name{aggregate_table_plot_features}
\alias{aggregate_table_plot_features}
\title{Aggregate table-referenced features}
\usage{
aggregate_table_plot_features(data, con)
}
\description{
Aggregate table-referenced features
}
\keyword{internal}
