% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_manip_db.R
\name{aggregate_traits_to_genus}
\alias{aggregate_traits_to_genus}
\title{Aggregate traits to genus level}
\usage{
aggregate_traits_to_genus(individuals, wd_fam_level)
}
\description{
Aggregate traits to genus level
}
\keyword{internal}
