% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_divid_plot.R
\name{approximate_isolated_xy}
\alias{approximate_isolated_xy}
\title{Interpolate x y position based on neighnour}
\usage{
approximate_isolated_xy(
  dataset,
  col_subplot = "quadrat",
  col_plot = "plot_name",
  col_pos_x = "position_x_iphone",
  col_pos_y = "position_y_iphone",
  tag = "tag"
)
}
\arguments{
\item{dataset}{extract}

\item{col_subplot}{string column name of quadrat}

\item{col_plot}{string column name of plot}

\item{col_pos_x}{numeric column name of x relative positioning within quadrat}

\item{col_pos_y}{numeric column name of y relative positioning within quadrat}

\item{tag}{numeric tag of individual within plot}
}
\description{
Experimental - Interpolate x y position based on neighnour, for isolated missing value only
}
\details{
Interpolate and fill x and y positioning
}
\author{
Gilles Dauby
}
