% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections_db.R
\name{check_taxa_permissions}
\alias{check_taxa_permissions}
\title{Check taxa database permissions}
\usage{
check_taxa_permissions(con)
}
\arguments{
\item{con}{A database connection object.}
}
\value{
Called for side effects. Prints messages about database access permissions
and warns if write access is detected.
}
\description{
A short description...
}
