% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_manip_db.R
\name{choose_growth_form}
\alias{choose_growth_form}
\title{Choose growth forms}
\usage{
choose_growth_form()
}
\value{
A tibble
}
\description{
Return a tibble of growth form chosen by hierarchy
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
