% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{choose_prompt}
\alias{choose_prompt}
\title{Choose from prompt}
\usage{
choose_prompt(choices_vec = c("Yes", "No", "Cancel"), message = "")
}
\arguments{
\item{choices_vec}{Optional. A character vector of choices.}

\item{message}{Optional. A single string for the prompt message.}
}
\value{
A logical value: `TRUE` for the first choice, `FALSE` for the second choice,
and `NA` for the third choice. For invalid selections, the function will
print an error message but may return an unexpected value.
}
\description{
Show prompt in console to choose which directions to take
}
