% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections_db.R
\name{connect_database}
\alias{connect_database}
\title{Connect to database}
\usage{
connect_database(
  db_type = c("main", "taxa"),
  pass = NULL,
  user = NULL,
  reset = FALSE,
  retry = TRUE
)
}
\arguments{
\item{db_type}{One of `"main"` or `"taxa"`.}

\item{pass}{A single string. Optional.}

\item{user}{A single string. Optional.}

\item{reset}{A single logical value. Optional.}

\item{retry}{A single logical value. Optional.}
}
\value{
A database connection object. The function will error if the connection
fails after the maximum number of attempts.
}
\description{
A short description...
}
