% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_features_function.R
\name{count_measurement_features}
\alias{count_measurement_features}
\title{Count measurement features}
\usage{
count_measurement_features(id_trait_measures, config, con)
}
\description{
Count measurement features
}
\keyword{internal}
