% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_manip_db.R
\name{country_list}
\alias{country_list}
\title{List of countries}
\usage{
country_list()
}
\value{
A tibble of all countries from table_countries
}
\description{
Provide list of valid countries from lookup table
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
