% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections_db.R
\name{define_read_only_policy}
\alias{define_read_only_policy}
\title{Helper functions for common policy scenarios}
\usage{
define_read_only_policy(con, user, ids, table = "data_liste_plots")
}
\description{
Helper functions for common policy scenarios
}
