% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections_db.R
\name{define_user_policy}
\alias{define_user_policy}
\title{Define user policy for row-level security}
\usage{
define_user_policy(
  con,
  user,
  ids,
  table = "data_liste_plots",
  policy_name = NULL,
  operations = "SELECT",
  drop_existing = TRUE
)
}
\description{
Define user policy for row-level security
}
