% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_divid_plot.R
\name{divid_plot}
\alias{divid_plot}
\title{Divid into quadrats a 1ha plot}
\usage{
divid_plot(corners)
}
\arguments{
\item{corners}{sf POINT the outputs WGS 84 of extract_corners function}

\item{plot_name}{vector string the name of the plot}
}
\value{
list sub_plot element is sf POLYGON with each quadrat as polygon
}
\description{
Divid into quadrats a 1ha plot
}
\author{
Hugo Leblanc
}
