% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomic_matching.R
\name{.add_synonym_info_sql}
\alias{.add_synonym_info_sql}
\title{Add synonym information to matches using SQL}
\usage{
.add_synonym_info_sql(matches, con)
}
\description{
Add synonym information to matches using SQL
}
\keyword{internal}
