% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomic_query_functions.R
\name{.add_traits_to_taxa}
\alias{.add_traits_to_taxa}
\title{Add traits to taxa (internal helper)}
\usage{
.add_traits_to_taxa(res)
}
\description{
Add traits to taxa (internal helper)
}
\keyword{internal}
