% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_manip_db.R
\name{.call.mydb}
\alias{.call.mydb}
\title{Load the database}
\usage{
.call.mydb()
}
\value{
The database is loaded
}
\description{
Load the database and ask for password
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
