% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomic_query_functions.R
\name{.clean_taxa_columns}
\alias{.clean_taxa_columns}
\title{Clean unwanted columns (internal helper)}
\usage{
.clean_taxa_columns(res)
}
\description{
Clean unwanted columns (internal helper)
}
\keyword{internal}
