% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_manip_db.R
\name{.comp_print_vec}
\alias{.comp_print_vec}
\title{Internal function}
\usage{
.comp_print_vec(vec_1, vec_2)
}
\arguments{
\item{vec_1}{tibble one row and same col numbers of vec_2}

\item{vec_2}{tibble one row and same col numbers of vec_1}
}
\value{
A list with one tibble of logical and a html table
}
\description{
Compare two rows and provide cols that differ and html with coloring different values
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
