% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_individuals_validation.R
\name{.compile_validation_messages}
\alias{.compile_validation_messages}
\title{Compile Validation Messages (Internal)}
\usage{
.compile_validation_messages(messages, severity)
}
\arguments{
\item{messages}{List of message strings}

\item{severity}{"error" or "warning"}
}
\value{
Data frame
}
\description{
Convert list of messages to data frame.
}
\keyword{internal}
