% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_functions.R
\name{.delete_colnam}
\alias{.delete_colnam}
\title{Delete an entry in colnam table}
\usage{
.delete_colnam(id)
}
\arguments{
\item{id}{integer}
}
\value{
No values
}
\description{
Delete an entry in colnam table using id for selection
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
