% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_manip_db.R
\name{.delete_entry_diconame}
\alias{.delete_entry_diconame}
\title{Delete an entry in trait measurement table}
\usage{
.delete_entry_diconame(id)
}
\arguments{
\item{id}{integer}
}
\value{
No values
}
\description{
Delete an entry in diconame table using id for selection
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
