% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_functions.R
\name{.delete_entry_sp_trait_measure}
\alias{.delete_entry_sp_trait_measure}
\title{Delete an entry in traits measurements table}
\usage{
.delete_entry_sp_trait_measure(id)
}
\arguments{
\item{id}{integer}
}
\value{
No values
}
\description{
Delete an entry in traits measurements table using id for selection
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
