% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_functions.R
\name{.delete_entry_trait_measure_features}
\alias{.delete_entry_trait_measure_features}
\title{Delete an entry in trait measurement features table}
\usage{
.delete_entry_trait_measure_features(id)
}
\arguments{
\item{id}{integer}
}
\value{
No values
}
\description{
Delete an entry in trait measurement features table using id for selection
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
