% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_features_function.R
\name{.delete_individual_feature_type}
\alias{.delete_individual_feature_type}
\title{Delete an entry in individual feature table}
\usage{
.delete_individual_feature_type(id)
}
\arguments{
\item{id}{integer}
}
\value{
No values
}
\description{
Delete an entry in individual/trait feature table using id for selection
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
