% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_functions.R
\name{.delete_individuals}
\alias{.delete_individuals}
\title{Delete an entry in individuals table}
\usage{
.delete_individuals(id)
}
\arguments{
\item{id}{integer}
}
\value{
No values
}
\description{
Delete an entry in individuals table using id for selection
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
