% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_manip_db.R
\name{.delete_plot}
\alias{.delete_plot}
\title{Delete an entry in plot meta-data}
\usage{
.delete_plot(id)
}
\arguments{
\item{id}{integer}
}
\value{
No values
}
\description{
Delete an entry in plot meta-data table using id for selection
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
