% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_functions.R
\name{.delete_specimens}
\alias{.delete_specimens}
\title{Delete an entry in specimen table}
\usage{
.delete_specimens(id)
}
\arguments{
\item{id}{integer}
}
\value{
No values
}
\description{
Delete an entry in specimen table using id for selection
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
