% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_manip_db.R
\name{.delete_subplotfeature}
\alias{.delete_subplotfeature}
\title{Delete an entry in subplotype_list table}
\usage{
.delete_subplotfeature(id)
}
\arguments{
\item{id}{integer}
}
\value{
No values
}
\description{
Delete an entry in subplotype_list table using id for selection
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
