% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_functions.R
\name{.delete_subplotfeature_feat}
\alias{.delete_subplotfeature_feat}
\title{Delete an entry in data_liste_sub_plots features}
\usage{
.delete_subplotfeature_feat(id)
}
\arguments{
\item{id}{integer}
}
\value{
No values
}
\description{
Delete an entry in data_liste_sub_plots features using id for selection
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
