% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_manip_db.R
\name{.delete_taxa}
\alias{.delete_taxa}
\title{Delete an entry in taxonomic table}
\usage{
.delete_taxa(id)
}
\arguments{
\item{id}{integer}
}
\value{
No values
}
\description{
Delete an entry in taxonomic table using id for selection
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
