% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_functions.R
\name{.delete_trait_list}
\alias{.delete_trait_list}
\title{Delete an entry in trait list}
\usage{
.delete_trait_list(id)
}
\arguments{
\item{id}{integer}
}
\value{
No values
}
\description{
Delete an entry in traitlist entry using id for selection
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
