% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updates_tables_functions.R
\name{.display_change}
\alias{.display_change}
\title{Display a single change with visual comparison}
\usage{
.display_change(id, column, old_value, new_value, id_column)
}
\arguments{
\item{id}{Record ID}

\item{column}{Column name}

\item{old_value}{Old value}

\item{new_value}{New value}

\item{id_column}{Name of ID column}
}
\description{
Display a single change with visual comparison
}
\keyword{internal}
