% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updates_tables_functions.R
\name{.display_changes_grouped}
\alias{.display_changes_grouped}
\title{Display multiple changes grouped by column}
\usage{
.display_changes_grouped(changes_df, id_column, max_display = 10)
}
\arguments{
\item{changes_df}{Dataframe with columns: id_column_value, column, old_value, new_value}

\item{id_column}{Name of ID column}

\item{max_display}{Maximum number of changes to display per column}
}
\description{
Display multiple changes grouped by column
}
\keyword{internal}
