% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_manip_db.R
\name{.find_ids}
\alias{.find_ids}
\title{Internal function}
\usage{
.find_ids(dataset, col_new, id_col_nbr, type_data)
}
\arguments{
\item{dataset}{tibble contain values to compare and id for matching}

\item{col_new}{string vector containing column names of dataset}

\item{id_col_nbr}{string vector}

\item{type_data}{string indicate which table of database is targetted. e.g. 'individuals'}
}
\value{
tibble
}
\description{
Compare values between of given columns and identify different values based on id matches
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
