% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_table_functions.R
\name{.find_similar_string}
\alias{.find_similar_string}
\title{Internal function}
\usage{
.find_similar_string(input, compared_table, column_name)
}
\arguments{
\item{input}{string vector of one value containing the string to compare}

\item{compared_table}{tibble including one column containing the different strings to which input should be compared}

\item{column_name}{string the column name of compared_table containing the compared values}
}
\value{
tibble
}
\description{
Looking for similar name
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
