% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomic_query_functions.R
\name{.format_taxa_names}
\alias{.format_taxa_names}
\title{Format taxonomic names (internal helper)}
\usage{
.format_taxa_names(res, mydb_taxa)
}
\description{
Format taxonomic names (internal helper)
}
\keyword{internal}
