% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_with_transactions.R
\name{.generate_admin_access_code}
\alias{.generate_admin_access_code}
\title{Generate Admin Access Code}
\usage{
.generate_admin_access_code(username, plot_names)
}
\arguments{
\item{username}{Username who imported the plots}

\item{plot_names}{Vector of plot names}
}
\value{
Character string with R code
}
\description{
Generates R code for admin to grant row-level security access.
Uses plot_names (which user knows) instead of plot_ids (which user can't see).
}
\keyword{internal}
