% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_features_function.R
\name{.get_trait_individuals_values}
\alias{.get_trait_individuals_values}
\title{Internal function}
\usage{
.get_trait_individuals_values(
  traits,
  src_individuals = NULL,
  ids_plot = NULL,
  skip_dates = TRUE,
  show_multiple_census = FALSE,
  remove_obs_with_issue = TRUE
)
}
\arguments{
\item{traits}{string vector with trait needed}

\item{ids_plot}{numeric vector with id_plots requested}

\item{skip_dates}{logical whether include day, month and year of observations}

\item{remove_obs_with_issue}{logical}

\item{id_individuals}{numeric vector with id_n individuals requested}

\item{show_multiple_measures}{logical whether multiple measures (i.e. census or sometimes more than one value for given measure)}

\item{collapse_multiple_val}{logical whether multiple traits measures should be collapsed (resulting values as character, separated by dash)}
}
\value{
tibble
}
\description{
Get for each trait, a tibble of individuals with measures or observations, deal with several observations
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
