% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_manip_db.R
\name{.link_colnam}
\alias{.link_colnam}
\title{Internal function}
\usage{
.link_colnam(data_stand, collector_field, id_colnam = "id_colnam")
}
\arguments{
\item{data_stand}{string tibble}

\item{collector_field}{string vector}
}
\value{
vector
}
\description{
Compute
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
