% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_manip_db.R
\name{.link_country}
\alias{.link_country}
\title{Internal function}
\usage{
.link_country(data_stand, country_field, country_id_field = "id_country")
}
\arguments{
\item{data_stand}{tibble}

\item{country_field}{string vector}

\item{country_id_field}{name of the column that will store the id}
}
\value{
vector
}
\description{
Compute
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
