% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_manip_db.R
\name{.link_method}
\alias{.link_method}
\title{Internal function}
\usage{
.link_method(method)
}
\arguments{
\item{data_stand}{tibble}

\item{trait}{string vector}
}
\value{
vector
}
\description{
Compute
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
