% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_with_transactions.R
\name{.link_method_for_import}
\alias{.link_method_for_import}
\title{Link Method for Import}
\usage{
.link_method_for_import(data, con, interactive, dry_run, progress)
}
\description{
Uses existing .link_table() to match methods interactively.
}
\keyword{internal}
