% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_manip_db.R
\name{.link_plot_name}
\alias{.link_plot_name}
\title{Internal function}
\usage{
.link_plot_name(data_stand, plot_name_field)
}
\arguments{
\item{data_stand}{string tibble}

\item{plot_name_field}{string vector}
}
\value{
vector
}
\description{
Compute
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
