% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_table_functions.R
\name{.link_sp_trait}
\alias{.link_sp_trait}
\title{Internal function}
\usage{
.link_sp_trait(data_stand, trait)
}
\arguments{
\item{data_stand}{tibble}

\item{trait}{string vector}
}
\value{
vector
}
\description{
Compute
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
