% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_manip_db.R
\name{.link_subplotype}
\alias{.link_subplotype}
\title{Internal function}
\usage{
.link_subplotype(data_stand, subplotype)
}
\arguments{
\item{data_stand}{tibble}

\item{subplotype}{string vector}
}
\value{
vector
}
\description{
Compute
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
