% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomic_matching.R
\name{.match_exact_sql}
\alias{.match_exact_sql}
\title{Exact matching helper using SQL}
\usage{
.match_exact_sql(parsed, con, include_authors, max_matches)
}
\description{
Exact matching helper using SQL
}
\keyword{internal}
