% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomic_query_functions.R
\name{.match_taxonomic_level}
\alias{.match_taxonomic_level}
\title{Match taxonomic level (order/family/genus) using new matching functions}
\usage{
.match_taxonomic_level(
  names,
  level,
  field,
  exact_match,
  min_similarity,
  mydb_taxa,
  verbose
)
}
\description{
Match taxonomic level (order/family/genus) using new matching functions
}
\keyword{internal}
