% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_trimble_function.R
\name{.pairwise_string_similarity}
\alias{.pairwise_string_similarity}
\title{Internal function}
\usage{
.pairwise_string_similarity(string_vector)
}
\arguments{
\item{string_vector}{string vector}
}
\value{
vector
}
\description{
Compute pairwise string similarity
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
