% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_individuals_with_transactions.R
\name{.prepare_individuals_data}
\alias{.prepare_individuals_data}
\title{Prepare Individuals Data for Insert (Internal)}
\usage{
.prepare_individuals_data(individuals_data, progress = TRUE)
}
\arguments{
\item{individuals_data}{Data frame with individual data and id_liste_plots}

\item{progress}{Show progress}
}
\value{
Data frame ready for insert
}
\description{
Prepares data frame for insertion into data_individuals table.
Selects and orders columns appropriately.
}
\keyword{internal}
