% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_with_transactions.R
\name{.prepare_plot_data}
\alias{.prepare_plot_data}
\title{Prepare Plot Data for data_liste_plots}
\usage{
.prepare_plot_data(data, people_columns, progress)
}
\description{
Removes people columns and adds modification dates.
}
\keyword{internal}
