% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomic_query_functions.R
\name{.print_taxa_results}
\alias{.print_taxa_results}
\title{Print taxa results (internal helper)}
\usage{
.print_taxa_results(res)
}
\description{
Print taxa results (internal helper)
}
\keyword{internal}
