% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomic_query_functions.R
\name{.query_taxa_by_ids}
\alias{.query_taxa_by_ids}
\title{Query taxa by IDs (internal helper)}
\usage{
.query_taxa_by_ids(
  ids,
  class,
  mydb_taxa,
  only_genus,
  only_family,
  only_class,
  check_synonymy,
  extract_traits,
  verbose
)
}
\description{
Query taxa by IDs (internal helper)
}
\keyword{internal}
