% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.rename_data}
\alias{.rename_data}
\title{Internal function}
\usage{
.rename_data(dataset, col_old, col_new)
}
\arguments{
\item{dataset}{tibble}

\item{col_old}{string vector}

\item{col_new}{string vector}
}
\value{
tibble
}
\description{
rename columns based on new and old columns names
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
Rename columns in a dataset
}
