% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_validation.R
\name{.resolve_lookup_column}
\alias{.resolve_lookup_column}
\title{Resolve Lookup Column with Interactive Matching}
\usage{
.resolve_lookup_column(
  data,
  column_name,
  lookup_function,
  lookup_value_col,
  lookup_id_col,
  lookup_table_name,
  con,
  interactive,
  fix_on_fly
)
}
\description{
Helper function to validate and interactively fix a lookup column.
}
\keyword{internal}
