% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomic_query_functions.R
\name{.resolve_synonyms}
\alias{.resolve_synonyms}
\title{Resolve synonyms (internal helper)}
\usage{
.resolve_synonyms(res, mydb_taxa, verbose)
}
\description{
Resolve synonyms (internal helper)
}
\keyword{internal}
