% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_individuals_column_mapping.R
\name{.separate_individuals_features}
\alias{.separate_individuals_features}
\title{Separate Flat Table into Individuals and Features - Internal}
\usage{
.separate_individuals_features(data, column_classifications)
}
\arguments{
\item{data}{Original flat table data frame}

\item{column_classifications}{Data frame with classification results}
}
\value{
List with individuals and features dataframes
}
\description{
Based on column classifications, separates the flat table into
two dataframes: individuals (core data) and features (traits).
}
\keyword{internal}
