% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_manip_db.R
\name{.split_censuses}
\alias{.split_censuses}
\title{Split plot data into census}
\usage{
.split_censuses(meta, dataset)
}
\arguments{
\item{meta}{tibble output of query_plot with no export individuals}

\item{dataset}{tibble output of query_plot with export individuals}
}
\value{
A list with as many tibble as census
}
\description{
split plot data into a list where each element is a census
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
