% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_manip_db.R
\name{.time_diff}
\alias{.time_diff}
\title{Add time difference in number of days for two census}
\usage{
.time_diff(census1, census2)
}
\arguments{
\item{census1}{tibble with first census}

\item{census2}{tibble with second census}
}
\value{
tibble joining both census and with a added column indicating time intervall
}
\description{
Add
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
