% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_manip_db.R
\name{.trim.growth}
\alias{.trim.growth}
\title{Identify potential errors for estimating growth}
\usage{
.trim.growth(
  censuses,
  slope = 0.006214,
  intercept = 0.9036,
  err.limit = 4,
  maxgrow = 75,
  mindbh = 100
)
}
\arguments{
\item{censuses}{tibble with first census}

\item{slope}{numeric see http://ctfs.si.edu/Public/CTFSRPackage/index.php/web/topics/growth~slash~growth.r/trim.growth}

\item{intercept}{numeric see http://ctfs.si.edu/Public/CTFSRPackage/index.php/web/topics/growth~slash~growth.r/trim.growth}

\item{err.limit}{integer any measure of second diameter higher than err.limit standard deviation below the first measure will be excluded}

\item{maxgrow}{numeric any growth in mm/year higher than maxgrow will be excluded}

\item{mindbh}{numeric minimum diameter in mm for excluding measures}
}
\value{
tibble joining both census and with a added column indicating logical whether inidividual should be excluded
}
\description{
Add a column of individuals to be excluded because of potential errors
Adapted from http://ctfs.si.edu/Public/CTFSRPackage/
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
