% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_individuals_validation.R
\name{.validate_feature_individual_linkage}
\alias{.validate_feature_individual_linkage}
\title{Validate Feature-Individual Linkage (Internal)}
\usage{
.validate_feature_individual_linkage(features_data, individuals_data)
}
\arguments{
\item{features_data}{Features data frame}

\item{individuals_data}{Individuals data frame}
}
\value{
List with errors and warnings
}
\description{
Check that features link to individuals in import.
}
\keyword{internal}
