% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_individuals_validation.R
\name{.validate_feature_linking_columns}
\alias{.validate_feature_linking_columns}
\title{Validate Feature Linking Columns (Internal)}
\usage{
.validate_feature_linking_columns(features_data)
}
\arguments{
\item{features_data}{Features data frame}
}
\value{
List of error messages
}
\description{
Check that features sheet has required linking columns.
}
\keyword{internal}
