% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_individuals_validation.R
\name{.validate_plot_access}
\alias{.validate_plot_access}
\title{Validate Plot Access (Internal)}
\usage{
.validate_plot_access(data, con)
}
\arguments{
\item{data}{Data frame with plot_name column}

\item{con}{Database connection}
}
\value{
List with errors and warnings
}
\description{
Check that plots exist in database and user has access.
}
\keyword{internal}
