% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_individuals_validation.R
\name{.validate_required_fields_individuals}
\alias{.validate_required_fields_individuals}
\title{Validate Required Fields for Individuals (Internal)}
\usage{
.validate_required_fields_individuals(data, required_cols)
}
\arguments{
\item{data}{Data frame}

\item{required_cols}{Required column names}
}
\value{
List of error messages
}
\description{
Validate Required Fields for Individuals (Internal)
}
\keyword{internal}
