% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_features_function.R
\name{enrich_census_info}
\alias{enrich_census_info}
\title{Enrich data with census information}
\usage{
enrich_census_info(data, con)
}
\description{
Enrich data with census information
}
\keyword{internal}
