% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_features_function.R
\name{enrich_measurement_features}
\alias{enrich_measurement_features}
\title{Enrich data with measurement-level features/metadata}
\usage{
enrich_measurement_features(data, con, src = "individuals")
}
\description{
Enrich data with measurement-level features/metadata
}
\keyword{internal}
