% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsplots_features_function.R
\name{enrich_subplot_obs_features}
\alias{enrich_subplot_obs_features}
\title{Enrich with subplot observation features}
\usage{
enrich_subplot_obs_features(data, con)
}
\description{
Enrich with subplot observation features
}
\keyword{internal}
