% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_traits_function.R
\name{enrich_with_taxa_info}
\alias{enrich_with_taxa_info}
\title{Enrich trait data with taxonomic information}
\usage{
enrich_with_taxa_info(data, con)
}
\description{
Enrich trait data with taxonomic information
}
\keyword{internal}
