% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_manip_db.R
\name{explore_allometric_taxa}
\alias{explore_allometric_taxa}
\title{Explore allometric relation}
\usage{
explore_allometric_taxa(
  genus_searched = NULL,
  tax_esp_searched = NULL,
  tax_fam_searched = NULL,
  id_search = NULL
)
}
\arguments{
\item{genus_searched}{string}

\item{tax_esp_searched}{string}

\item{tax_fam_searched}{string}

\item{id_search}{integer}
}
\value{
A tibble

A tibble of taxa or individuals if extract_individuals is TRUE
}
\description{
Provide allometric data and graph dbh-height of selected taxa
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
