% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsplots_features_function.R
\name{extract_census_dates}
\alias{extract_census_dates}
\title{Extract and format census dates}
\usage{
extract_census_dates(census_data)
}
\description{
Extract and format census dates
}
\keyword{internal}
