% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_divid_plot.R
\name{extract_corners}
\alias{extract_corners}
\title{Extract all corners of 1ha plot}
\usage{
extract_corners(coordinates, map_res = FALSE)
}
\arguments{
\item{coordinates}{tibble output of query_plots, element coordinates}
}
\value{
list
}
\description{
Extract all corners in UTM and geographic coordinates of 1ha plot and map it as well
}
\author{
Hugo Leblanc
}
