% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_features_function.R
\name{fetch_linked_individuals}
\alias{fetch_linked_individuals}
\title{Fetch linked individual data}
\usage{
fetch_linked_individuals(individual_ids, con, chunk_size = 30000)
}
\description{
Fetch linked individual data
}
\keyword{internal}
