% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsplots_features_function.R
\name{fetch_subplot_features}
\alias{fetch_subplot_features}
\title{Fetch subplot features from database}
\usage{
fetch_subplot_features(plot_ids, subplot_ids, con)
}
\description{
Fetch subplot features from database
}
\keyword{internal}
