% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsplots_features_function.R
\name{fetch_subplot_obs_features}
\alias{fetch_subplot_obs_features}
\title{Fetch subplot observation features}
\usage{
fetch_subplot_obs_features(subplot_ids, con)
}
\description{
Fetch subplot observation features
}
\keyword{internal}
