% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_traits_function.R
\name{fetch_taxa_trait_measurements}
\alias{fetch_taxa_trait_measurements}
\title{Fetch raw trait measurements for taxa}
\usage{
fetch_taxa_trait_measurements(idtax, trait_ids = NULL, con)
}
\description{
Fetch raw trait measurements for taxa
}
\keyword{internal}
