% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_features_function.R
\name{fetch_with_chunking}
\alias{fetch_with_chunking}
\title{Generic chunking function for large queries}
\usage{
fetch_with_chunking(ids, query_fun, chunk_size, con, desc = "data")
}
\description{
Generic chunking function for large queries
}
\keyword{internal}
