% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_manip_db.R
\name{func_try_fetch}
\alias{func_try_fetch}
\title{Send query to postgresql database}
\usage{
func_try_fetch(con, sql)
}
\description{
Access to postgresql database with repeating if no successful
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
