% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_structure.R
\name{get_database_fk}
\alias{get_database_fk}
\title{Get database foreign keys}
\usage{
get_database_fk(con)
}
\arguments{
\item{con}{A database connection. Optional; if `NULL`, will call `call.mydb()`.}
}
\value{
A database structure diagram showing foreign key relationships between
the specified tables.
}
\description{
Get a figure showing main tables and primary and foreign keys
}
