% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_features_function.R
\name{get_lookup_table_info}
\alias{get_lookup_table_info}
\title{Get lookup table information}
\usage{
get_lookup_table_info(table_name)
}
\description{
Get lookup table information
}
\keyword{internal}
