% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_features_function.R
\name{get_measurement_features_config}
\alias{get_measurement_features_config}
\title{Get configuration for measurement features}
\usage{
get_measurement_features_config(src)
}
\description{
Get configuration for measurement features
}
\keyword{internal}
