% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_divid_plot.R
\name{get_plot_rel_xy}
\alias{get_plot_rel_xy}
\title{Project stems in geographical space}
\usage{
get_plot_rel_xy(dataset)
}
\arguments{
\item{dataset}{extract of query_plots individuals TRUE}
}
\description{
Project stems in geogaphical space
}
\details{
Add columns x_100 and y_100 with relative position in plot based on position_x_iphone and positions_y_iphone
}
\author{
Gilles Dauby
}
