% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_manip_db.R
\name{get_ref_specimen_ind}
\alias{get_ref_specimen_ind}
\title{Find unlinked individual}
\usage{
get_ref_specimen_ind(collector = NULL, ids = NULL)
}
\arguments{
\item{ids}{vector of id_n of individuals for which we want to explore links}
}
\value{
list
}
\description{
Extract individuals for which no links exist with herbarium specimens
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
