% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_translations.R
\name{get_translations}
\alias{get_translations}
\title{Get translation dictionary}
\usage{
get_translations(language = "en")
}
\arguments{
\item{language}{Character, either "en" or "fr"}
}
\value{
Named list of translations
}
\description{
Get translation dictionary
}
\keyword{internal}
