% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_manip_db.R
\name{growth_computing}
\alias{growth_computing}
\title{Growth computing for multiple census}
\usage{
growth_computing(
  dataset,
  metadata,
  mindbh = NULL,
  err.limit = 4,
  maxgrow = 75,
  method = "I"
)
}
\arguments{
\item{dataset}{tibble, ouput of query_plots}

\item{metadata}{tibble, ouput of query_plots}

\item{mindbh}{numeric see http://ctfs.si.edu/Public/CTFSRPackage/index.php/web/topics/growth~slash~growth.r/trim.growth}

\item{err.limit}{integer any measure of second diameter higher than err.limit standard deviation below the first measure will be excluded}

\item{maxgrow}{numeric any growth in mm/year higher than maxgrow will be excluded}

\item{method}{string either 'I' or 'E'}
}
\value{
tibble
}
\description{
Growth computing for multiple census
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
