% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_manip_db.R
\name{herbarium_label}
\alias{herbarium_label}
\title{Herbarium labels}
\usage{
herbarium_label(
  dat = NULL,
  theme = c("GILLES"),
  outfile = "herblabel.rtf",
  font = c("Roman", "Arial"),
  font_size = 1
)
}
\arguments{
\item{dat}{data.frame}

\item{theme}{string}

\item{outfile}{string}

\item{font}{string}

\item{font_size}{integer}
}
\value{
A tibble
}
\description{
Produce rtf file with herbarium specimens labels
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
