% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_divid_plot.R
\name{latlong2UTM}
\alias{latlong2UTM}
\title{Get UTM from geographical coordinates}
\usage{
latlong2UTM(coord)
}
\arguments{
\item{coord}{sf POINT the outputs WGS 84 of extract_corners function}
}
\description{
Get UTM from geographical coordinates
}
\author{
Hugo Leblanc
}
