% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_manip_db.R
\name{launch_query_tax_app}
\alias{launch_query_tax_app}
\title{Launch shiny app for taxa}
\usage{
launch_query_tax_app()
}
\value{
Launch in web browser the app
}
\description{
Allow exploring the database by taxa - export individuals or taxa list and plot stem diameter and tree height
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
