% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_manip_db.R
\name{launch_stand_people_app}
\alias{launch_stand_people_app}
\title{Launch shiny app for standardizing people names}
\usage{
launch_stand_people_app()
}
\value{
Launch in web browser the app
}
\description{
Application for standardizing people names, especially specimens collectors
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
