% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_features_function.R
\name{list_individual_features}
\alias{list_individual_features}
\title{List all available individual features}
\usage{
list_individual_features(con = NULL)
}
\description{
List all available individual features
}
