% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_manip_db.R
\name{method_list}
\alias{method_list}
\title{List of method}
\usage{
method_list()
}
\value{
A tibble of all method
}
\description{
Provide list of method where plots occur
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
