% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_auto_matching.R
\name{mod_auto_matching_ui}
\alias{mod_auto_matching_ui}
\title{Auto Matching Module - UI}
\usage{
mod_auto_matching_ui(id)
}
\arguments{
\item{id}{Character, module ID}
}
\value{
Shiny UI element
}
\description{
Auto Matching Module - UI
}
\keyword{internal}
