% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_column_select.R
\name{mod_column_select_ui}
\alias{mod_column_select_ui}
\title{Column Select Module - UI}
\usage{
mod_column_select_ui(id)
}
\arguments{
\item{id}{Character, module ID}
}
\value{
Shiny UI element
}
\description{
Column Select Module - UI
}
\keyword{internal}
