% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_data_input.R
\name{mod_data_input_ui}
\alias{mod_data_input_ui}
\title{Data Input Module - UI}
\usage{
mod_data_input_ui(id)
}
\arguments{
\item{id}{Character, module ID}
}
\value{
Shiny UI element
}
\description{
Data Input Module - UI
}
\keyword{internal}
