% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_fuzzy_suggestions.R
\name{mod_fuzzy_suggestions_server}
\alias{mod_fuzzy_suggestions_server}
\title{Fuzzy Suggestions Module - Server}
\usage{
mod_fuzzy_suggestions_server(
  id,
  input_name,
  max_suggestions = shiny::reactive(10),
  min_similarity = shiny::reactive(0.3),
  include_authors = shiny::reactive(FALSE),
  i18n
)
}
\arguments{
\item{id}{Character, module ID}

\item{input_name}{Reactive character, the name to find suggestions for}

\item{max_suggestions}{Reactive or numeric, maximum suggestions to show}

\item{min_similarity}{Reactive or numeric, minimum similarity threshold}

\item{include_authors}{Reactive or logical, whether to include author names}

\item{i18n}{shiny.i18n Translator object}
}
\value{
Reactive integer, idtax_n of selected suggestion (or NULL)
}
\description{
Fuzzy Suggestions Module - Server
}
\keyword{internal}
