% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_fuzzy_suggestions.R
\name{mod_fuzzy_suggestions_ui}
\alias{mod_fuzzy_suggestions_ui}
\title{Fuzzy Suggestions Module - UI}
\usage{
mod_fuzzy_suggestions_ui(id)
}
\arguments{
\item{id}{Character, module ID}
}
\value{
Shiny UI element
}
\description{
Fuzzy Suggestions Module - UI
}
\keyword{internal}
