% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_language_toggle.R
\name{mod_language_toggle_server}
\alias{mod_language_toggle_server}
\title{Language Toggle Module - Server}
\usage{
mod_language_toggle_server(id, initial = "en")
}
\arguments{
\item{id}{Character, module ID}

\item{initial}{Character, initial language ("en" or "fr")}
}
\value{
Reactive value containing current language
}
\description{
Language Toggle Module - Server
}
\keyword{internal}
