% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_language_toggle.R
\name{mod_language_toggle_ui}
\alias{mod_language_toggle_ui}
\title{Language Toggle Module - UI}
\usage{
mod_language_toggle_ui(id)
}
\arguments{
\item{id}{Character, module ID}
}
\value{
Shiny UI element
}
\description{
Language Toggle Module - UI
}
\keyword{internal}
