% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_name_review.R
\name{mod_name_review_ui}
\alias{mod_name_review_ui}
\title{Name Review Module - UI}
\usage{
mod_name_review_ui(id)
}
\arguments{
\item{id}{Character, module ID}
}
\value{
Shiny UI element
}
\description{
Name Review Module - UI
}
\keyword{internal}
