% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_progress_tracker.R
\name{mod_progress_tracker_server}
\alias{mod_progress_tracker_server}
\title{Progress Tracker Module - Server}
\usage{
mod_progress_tracker_server(
  id,
  match_results,
  language = shiny::reactive("en")
)
}
\arguments{
\item{id}{Character, module ID}

\item{match_results}{Reactive list containing matching results}

\item{language}{Reactive returning current language ("en" or "fr")}
}
\value{
NULL (display only module)
}
\description{
Progress Tracker Module - Server
}
\keyword{internal}
