% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_progress_tracker.R
\name{mod_progress_tracker_ui}
\alias{mod_progress_tracker_ui}
\title{Progress Tracker Module - UI}
\usage{
mod_progress_tracker_ui(id)
}
\arguments{
\item{id}{Character, module ID}
}
\value{
Shiny UI element
}
\description{
Progress Tracker Module - UI
}
\keyword{internal}
