% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_results_export.R
\name{mod_results_export_server}
\alias{mod_results_export_server}
\title{Results Export Module - Server}
\usage{
mod_results_export_server(id, results, original_data, i18n)
}
\arguments{
\item{id}{Character, module ID}

\item{results}{Reactive list from auto matching module}

\item{original_data}{Reactive data.frame, original user data}

\item{i18n}{shiny.i18n Translator object}
}
\value{
NULL (handles download only)
}
\description{
Results Export Module - Server
}
\keyword{internal}
