% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_results_export.R
\name{mod_results_export_ui}
\alias{mod_results_export_ui}
\title{Results Export Module - UI}
\usage{
mod_results_export_ui(id)
}
\arguments{
\item{id}{Character, module ID}
}
\value{
Shiny UI element
}
\description{
Results Export Module - UI
}
\keyword{internal}
