% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_traits_enrichment.R
\name{mod_traits_enrichment_server}
\alias{mod_traits_enrichment_server}
\title{Traits Enrichment Module - Server}
\usage{
mod_traits_enrichment_server(id, results, language = shiny::reactive("en"))
}
\arguments{
\item{id}{Character, module ID}

\item{results}{Reactive list from review module (contains final matched data)}

\item{language}{Reactive returning current language ("en" or "fr")}
}
\description{
Traits Enrichment Module - Server
}
\keyword{internal}
