% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_traits_enrichment.R
\name{mod_traits_enrichment_ui}
\alias{mod_traits_enrichment_ui}
\title{Traits Enrichment Module - UI}
\usage{
mod_traits_enrichment_ui(id)
}
\arguments{
\item{id}{Character, module ID}
}
\value{
Shiny UI element
}
\description{
Traits Enrichment Module - UI
}
\keyword{internal}
