% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{model_wd_1}
\alias{model_wd_1}
\title{model_wd_1}
\format{
An object of class lm
}
\usage{
data(model_wd_2)
}
\description{
Linear model to predict whole-tree wood density from small branches wood density
}
\examples{
data(model_wd_2)
}
\references{
Momo et al 2020
}
\keyword{datasets}
