% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_manip_db.R
\docType{data}
\name{phylo_tree}
\alias{phylo_tree}
\title{phylo_tree}
\format{
An object of class phylo
}
\usage{
data(phylo_tree)
}
\description{
Phylogeny from Janssens et al 2020
}
\examples{
data(phylo_tree)
}
\keyword{datasets}
