% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_features_function.R
\name{pivot_character_traits}
\alias{pivot_character_traits}
\title{Pivot character traits}
\usage{
pivot_character_traits(data, include_census)
}
\description{
Pivot character traits
}
\keyword{internal}
