% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_features_function.R
\name{pivot_features_by_type}
\alias{pivot_features_by_type}
\title{Pivot features by type with custom aggregation}
\usage{
pivot_features_by_type(data, value_col, id_col, agg_fun)
}
\description{
Pivot features by type with custom aggregation
}
\keyword{internal}
