% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_features_function.R
\name{pivot_measurement_features}
\alias{pivot_measurement_features}
\title{Pivot measurement features to wide format}
\usage{
pivot_measurement_features(data, config)
}
\description{
Pivot measurement features to wide format
}
\keyword{internal}
