% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_features_function.R
\name{pivot_numeric_traits}
\alias{pivot_numeric_traits}
\title{Pivot numeric traits}
\usage{
pivot_numeric_traits(data, include_census)
}
\description{
Pivot numeric traits
}
\keyword{internal}
