% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_features_function.R
\name{pivot_table_references}
\alias{pivot_table_references}
\title{Pivot table-referenced features}
\usage{
pivot_table_references(data, id_col)
}
\description{
Pivot table-referenced features
}
\keyword{internal}
