% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_features_function.R
\name{pivot_traits_to_wide}
\alias{pivot_traits_to_wide}
\title{Pivot trait measurements to wide format}
\usage{
pivot_traits_to_wide(data, include_census = FALSE)
}
\description{
Pivot trait measurements to wide format
}
\keyword{internal}
