% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_styles_helpers.R
\name{print.plot_query_list}
\alias{print.plot_query_list}
\title{Print method for plot_query_list}
\usage{
\method{print}{plot_query_list}(x, ...)
}
\arguments{
\item{x}{A plot_query_list object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for plot_query_list
}
