% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_validation.R
\name{print.plot_validation_result}
\alias{print.plot_validation_result}
\title{Print method for validation results}
\usage{
\method{print}{plot_validation_result}(x, ...)
}
\arguments{
\item{x}{Validation result object}

\item{...}{Additional arguments passed to print_validation_results}
}
\description{
Print method for validation results
}
