% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_with_transactions.R
\name{print_import_result}
\alias{print_import_result}
\title{Print Import Result}
\usage{
print_import_result(result)
}
\arguments{
\item{result}{Import result from import_plot_metadata()}
}
\description{
Pretty-prints import results.
}
