% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_individuals_validation.R
\name{print_individual_validation_results}
\alias{print_individual_validation_results}
\title{Print Individual Validation Results}
\usage{
print_individual_validation_results(validation)
}
\arguments{
\item{validation}{Validation result from validate_individual_data()}
}
\description{
Pretty print individual data validation results.
}
