% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_column_mapping.R
\name{print_mapping_summary}
\alias{print_mapping_summary}
\title{Print Mapping Summary}
\usage{
print_mapping_summary(mapping_result)
}
\arguments{
\item{mapping_result}{Result from map_user_columns()}
}
\value{
Invisibly returns a summary data frame
}
\description{
Display detailed summary of column mappings
}
\examples{
\dontrun{
mapping <- map_user_columns(my_data, config)
print_mapping_summary(mapping)
}

}
