% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_manip_db.R
\name{print_table}
\alias{print_table}
\title{print table as html in viewer}
\usage{
print_table(res_print)
}
\arguments{
\item{res_print}{tibble}
}
\value{
print html in viewer
}
\description{
print table as html in viewer reordered
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
