% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_manip_db.R
\name{process_stems}
\alias{process_stems}
\title{Process multiple stems}
\usage{
process_stems(individuals, concatenate = FALSE)
}
\arguments{
\item{individuals}{Data frame of individuals}

\item{concatenate}{If TRUE, concatenate multiple stems}
}
\value{
Data frame with processed stems
}
\description{
Concatenate multiple stems if requested
}
