% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_manip_db.R
\name{process_trimble_data}
\alias{process_trimble_data}
\title{Process trimble data}
\usage{
process_trimble_data(PATH = NULL, plot_name = NULL, format = "dbf")
}
\arguments{
\item{PATH}{string path to trimble data}

\item{plot_name}{string plot name to give to plot}
}
\value{
A list with two tibbles
}
\description{
Process raw trimble data and provide individuals data ready to import and log file with potential problems
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
