% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_manip_db.R
\name{province_list}
\alias{province_list}
\title{List of province}
\usage{
province_list(country = NULL)
}
\arguments{
\item{country}{string}
}
\value{
A tibble of all province
}
\description{
Provide list of province where plots occur
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
