% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_manip_db.R
\name{query_colnam}
\alias{query_colnam}
\title{Query in colnam table}
\usage{
query_colnam(id_colnam = NULL, pattern = NULL)
}
\arguments{
\item{pattern}{string vector trait to look for in the table}

\item{id_trait}{integer id of trait to select}
}
\value{
tibble with query results
}
\description{
Query in colnam table by id or pattern
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
