% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{query_exact_match}
\alias{query_exact_match}
\title{Query exact match}
\usage{
query_exact_match(tbl, field, values_q, con)
}
\arguments{
\item{tbl}{tibble with one field listing names to be searched}

\item{field}{string column name to be search}

\item{values_q}{string names to be searched}

\item{con}{PqConnection connection to RPostgres database}
}
\value{
A list of two elements: (1) res_q with matched records, (2) query_tb with match status
}
\description{
`r lifecycle::badge("superseded")`

This function has been superseded by the intelligent matching functions in
`taxonomic_matching.R`. For better quality matches that handle infraspecific
ranks properly, use [match_taxonomic_names()] with `method = "exact"`.

Extract from a sql database an exact match on a given field.
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
