% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{query_fuzzy_match}
\alias{query_fuzzy_match}
\title{Query fuzzy match}
\usage{
query_fuzzy_match(tbl, field, values_q, con)
}
\arguments{
\item{tbl}{tibble with one field listing names to be searched}

\item{field}{string column name to be search}

\item{values_q}{string names to be searched}

\item{con}{PqConnection connection to RPostgres database}
}
\value{
A tibble with matched records from the database
}
\description{
`r lifecycle::badge("superseded")`

This function has been superseded by the intelligent matching functions in
`taxonomic_matching.R`. For better quality matches with genus-constrained
fuzzy search, use [match_taxonomic_names()] instead.

Extract from a sql database a fuzzy match on a given field using PostgreSQL SIMILARITY.
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
