% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specimen_linking_functions.R
\name{query_link_individual_specimen}
\alias{query_link_individual_specimen}
\title{Query link between specimens and individuals}
\usage{
query_link_individual_specimen(id_ind = NULL, id_specimen = NULL)
}
\arguments{
\item{id_ind}{integer}

\item{id_specimen}{integer}
}
\value{
No values
}
\description{
Query links between specimens and individuals
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
