% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_manip_db.R
\name{query_specimens}
\alias{query_specimens}
\title{Exploring specimens data}
\usage{
query_specimens(collector = NULL, id_colnam = NULL, number = NULL,
  number_min = NULL, number_max = NULL, genus_searched = NULL,
  tax_esp_searched = NULL, tax_fam_searched = NULL, id_search = NULL,
  subset_columns = TRUE, show_previous_modif = TRUE,
  generate_labels = FALSE,
  project_title = "Reference specimens collected in trees inventory",
  file_labels = "labels")
}
\arguments{
\item{collector}{string collector name}

\item{id_colnam}{integer id of collector}

\item{number}{integer specimen number}

\item{number_min}{integer minimum specimen number}

\item{number_max}{integer maximum specimen number}

\item{genus_searched}{string specimens of genus searched}

\item{tax_esp_searched}{string specimens of species searched}

\item{tax_fam_searched}{string specimens of family searched}

\item{id_search}{integer id searched}

\item{subset_columns}{logical if only a subset of columns should be provided}

\item{show_previous_modif}{logical if you want to see previous modification of the entry - useful to see previous identification for example}

\item{generate_labels}{logical if labels should be produced}

\item{project_title}{string if labels are produced title of the label}

\item{file_labels}{string if labels are produced name of the rtf file}
}
\value{
A tibble
}
\description{
Exploring specimens data and if necessary export labels
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
