% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsplots_features_function.R
\name{query_subplot_observations_feat}
\alias{query_subplot_observations_feat}
\title{List, selected trait measures features}
\usage{
query_subplot_observations_feat(id_sub_plots = NULL, pivot_table = TRUE)
}
\arguments{
\item{id_trait_measures}{integer}
}
\value{
A tibble of all subplots
}
\description{
Table of srait measures features
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
