% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_manip_db.R
\name{query_tax_all}
\alias{query_tax_all}
\title{Query taxa}
\usage{
query_tax_all(
  genus_searched = NULL,
  tax_esp_searched = NULL,
  tax_fam_searched = NULL,
  tax_name1_searched = NULL,
  id_search = NULL,
  show_synonymies = TRUE,
  extract_individuals = FALSE
)
}
\arguments{
\item{genus_searched}{string}

\item{tax_esp_searched}{string}

\item{tax_fam_searched}{string}

\item{tax_name1_searched}{string}

\item{id_search}{integer}

\item{show_synonymies}{logical}

\item{extract_individuals}{logical}
}
\value{
A tibble

A tibble of taxa or individuals if extract_individuals is TRUE
}
\description{
Provide information on selected taxa
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
