% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_manip_db.R
\name{query_taxa_app}
\alias{query_taxa_app}
\title{Launch shiny app for taxa}
\usage{
query_taxa_app()
}
\value{
Launch in web browser the app
}
\description{
Allow exploring the database by taxa
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
