% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_manip_db.R
\name{query_trait}
\alias{query_trait}
\title{Query in taxa trait table}
\usage{
query_trait(id_trait = NULL, pattern = NULL)
}
\arguments{
\item{id_trait}{integer id of trait to select}

\item{pattern}{string vector trait to look for in the table}
}
\value{
tibble with query results
}
\description{
Query in taxa trait table by id or pattern
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
