% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_manip_db.R
\name{reorganize_individual_columns}
\alias{reorganize_individual_columns}
\title{Reorganize columns for individuals data}
\usage{
reorganize_individual_columns(individuals)
}
\arguments{
\item{individuals}{Data frame of individuals}
}
\value{
Data frame with reorganized columns
}
\description{
Reorganize columns for individuals data
}
\keyword{internal}
