% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsplots_features_function.R
\name{resolve_table_references}
\alias{resolve_table_references}
\title{Resolve table references in features}
\usage{
resolve_table_references(data, con)
}
\description{
Resolve table references in features
}
\keyword{internal}
