% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections_db.R
\name{setup_db_credentials}
\alias{setup_db_credentials}
\title{Setup credentials storage in environment variables}
\usage{
setup_db_credentials(user = NULL, pass = NULL)
}
\arguments{
\item{user}{Username for database}

\item{pass}{Password for database}
}
\description{
Helper function to configure credentials in .Renviron file.
WARNING: Credentials will be stored in plain text. Only use on secure personal computers.
}
