% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_manip_db.R
\name{species_plot_matrix}
\alias{species_plot_matrix}
\title{Get species-plot data frame}
\usage{
species_plot_matrix(data_tb, tax_col = "tax_sp_level", plot_col = "plot_name")
}
\value{
A data frame with taxa as row, plot as columns and values as number of individuals
}
\description{
Convert the extract of database into a species-plot data frame
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
