% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsplots_features_function.R
\name{subplot_list}
\alias{subplot_list}
\title{List of subplot features}
\usage{
subplot_list()
}
\value{
A tibble of all subplot features
}
\description{
Provide list of subplot features of plot
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
