% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_features_function.R
\name{summarize_feature}
\alias{summarize_feature}
\title{Get summary statistics for a specific feature}
\usage{
summarize_feature(trait_name, con = NULL)
}
\description{
Get summary statistics for a specific feature
}
