% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_translations.R
\name{t_}
\alias{t_}
\title{Translate a key}
\usage{
t_(key, language = "en")
}
\arguments{
\item{key}{Character, translation key}

\item{language}{Character, either "en" or "fr"}
}
\value{
Translated string
}
\description{
Helper function to translate a single key in the current language
}
\keyword{internal}
