% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_manip_db.R
\docType{data}
\name{table_taxa_tb}
\alias{table_taxa_tb}
\title{Table taxa}
\format{
An object data.frame/tibble
}
\usage{
data(table_taxa_tb)
}
\description{
Rainbio taxonomic backbone
}
\examples{
data(table_taxa_tb)
}
\keyword{datasets}
