% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections_db.R
\name{test_connection}
\alias{test_connection}
\title{Test database connection}
\usage{
test_connection(con)
}
\arguments{
\item{con}{A database connection object.}
}
\value{
`TRUE` if the connection is valid and a test query succeeds, `FALSE` otherwise.
}
\description{
To test if the database connection if valid
}
