% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_plots_v2_base.R
\name{test_plot_filtering}
\alias{test_plot_filtering}
\title{Tester les nouvelles classes (pour développement)}
\usage{
test_plot_filtering()
}
\description{
Fonction pour tester PlotFilterBuilder et PlotFetcher
}
