% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_manip_db.R
\name{traits_list}
\alias{traits_list}
\title{List of trait}
\usage{
traits_list(id_trait = NULL)
}
\value{
A tibble of all traits
}
\description{
Provide list of traits available
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
