% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_manip_db.R
\name{try_open_postgres_table_mem}
\alias{try_open_postgres_table_mem}
\title{Memoised try_open_postgres_table function}
\usage{
try_open_postgres_table_mem(table, con)
}
\description{
Memoised 'try_open_postgres_table' function
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
