% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updates_tables_functions.R
\name{update_colnam}
\alias{update_colnam}
\title{Update colnam}
\usage{
update_colnam(
  colnam_searched = NULL,
  colnam_id = NULL,
  new_colnam = NULL,
  new_surname = NULL,
  new_family_name = NULL,
  new_nationality = NULL,
  ask_before_update = TRUE,
  add_backup = TRUE,
  show_results = TRUE
)
}
\arguments{
\item{ask_before_update}{logical TRUE by default, ask for confirmation before updating}

\item{add_backup}{logical TRUE by default, add backup of modified data}

\item{show_results}{logical TRUE by default, show the data that has been modified}

\item{trait_searched}{string genus name searched}

\item{trait_id}{string genus name searched}

\item{new_trait_name}{string new trait name}

\item{new_relatedterm}{string new relatedterm name}

\item{new_maxallowedvalue}{numeric new maxallowedvalue}

\item{new_minallowedvalue}{numeric new minallowedvalue}

\item{new_traitdescription}{string new traitdescription}

\item{new_expectedunit}{string new expectedunit}
}
\value{
No return value individuals updated
}
\description{
Update colnam table
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
