% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_manip_db.R
\name{update_dico_name}
\alias{update_dico_name}
\title{Update taxonomic data}
\usage{
update_dico_name(genus_searched = NULL, tax_esp_searched = NULL,
  tax_fam_searched = NULL, new_tax_gen = NULL, new_tax_esp = NULL,
  new_full_name_auth = NULL, new_tax_fam = NULL,
  new_tax_rank1 = NULL, new_tax_name1 = NULL, new_taxook = NULL,
  new_morphocat = NULL, new_detvalue = NULL,
  new_full_name_no_auth = NULL, new_full_name_used = NULL,
  new_full_name_used2 = NULL, new_id_diconame_good = NULL,
  id_search = NULL, ask_before_update = TRUE, add_backup = TRUE,
  show_results = TRUE, no_synonym_modif = FALSE, synonym_of = NULL)
}
\arguments{
\item{genus_searched}{string genus name searched}

\item{tax_esp_searched}{string species name searched}

\item{tax_fam_searched}{string family name searched}

\item{new_tax_gen}{string new genus name}

\item{new_tax_esp}{string new species name}

\item{new_full_name_auth}{string new full name with authors}

\item{new_tax_fam}{string new family name}

\item{new_tax_rank1}{string new rank}

\item{new_tax_name1}{string new name of rank1}

\item{new_taxook}{integer new tax code}

\item{new_morphocat}{integer new morphocat code}

\item{new_detvalue}{integer new detvalue code}

\item{new_full_name_no_auth}{string new full name without authors - if not provided concatenate of new_esp and new_genus}

\item{new_full_name_used}{string new full_name_used}

\item{new_full_name_used2}{string new full_name_used2}

\item{new_id_diconame_good}{integer if the selected name should be put in synonymy, id of the taxa}

\item{id_search}{integer id of the taxa searched}

\item{ask_before_update}{logical TRUE by default, ask for confirmation before updating}

\item{add_backup}{logical TRUE by default, add backup of modified data}

\item{show_results}{logical TRUE by default, show the data that has been modified}

\item{no_synonym_modif}{logical FALSE by default, if TRUE and if the selected taxa is considered as synonym, then this will be modified and the selected taxa will not longer be a synonym}

\item{synonym_of}{list if the selected taxa should be put in synonymy with an existing taxa, add in a list at least one values to identify to which taxa it will be put in synonymy: genus, species or id}
}
\value{
No return value individuals updated
}
\description{
Update taxonomic data
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
