% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_manip_db.R
\name{update_dico_name_batch}
\alias{update_dico_name_batch}
\title{Update diconame data based on id of taxa}
\usage{
update_dico_name_batch(
  new_data,
  col_names_select,
  col_names_corresp,
  id_col,
  launch_update = FALSE,
  add_backup = TRUE,
  ask_before_update = FALSE,
  only_new_ident = FALSE
)
}
\arguments{
\item{new_data}{tibble}

\item{col_names_select}{vector string of new_data to be used for update}

\item{col_names_corresp}{vector string of corresponding fields to update}

\item{id_col}{integer indicate which name of col_names_select is the id for matching data}

\item{launch_update}{logical if TRUE updates are performed}

\item{add_backup}{logical whether backup of modified data should be recorded}
}
\value{
No return value individuals updated
}
\description{
Update diconame _ one or more entry at a time
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
