% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_manip_db.R
\name{update_ident_specimens}
\alias{update_ident_specimens}
\title{Update specimens table}
\usage{
update_ident_specimens(
  colnam = NULL,
  number = NULL,
  id_speci = NULL,
  new_genus = NULL,
  new_species = NULL,
  new_family = NULL,
  id_new_taxa = NULL,
  new_detd = NULL,
  new_detm = NULL,
  new_dety = NULL,
  new_detby = NULL,
  new_detvalue = NULL,
  add_backup = TRUE,
  show_results = TRUE,
  only_new_ident = TRUE,
  ask_before_update = TRUE
)
}
\arguments{
\item{colnam}{string collector name}

\item{number}{integer specimen number}

\item{id_speci}{integer id of specimen}

\item{new_genus}{string new genus name}

\item{new_species}{string new species name}

\item{new_family}{string new family name}

\item{id_new_taxa}{integer id of the new taxa}

\item{new_detd}{integer day of identification}

\item{new_detm}{logical if you want to see previous modification of the entry - useful to see previous identification for example}

\item{new_dety}{logical if labels should be produced}

\item{new_detby}{string if labels are produced title of the label}

\item{new_detvalue}{string if labels are produced name of the rtf file}

\item{add_backup}{string if labels are produced name of the rtf file}

\item{show_results}{string if labels are produced name of the rtf file}

\item{only_new_ident}{string if labels are produced name of the rtf file}
}
\value{
A tibble
}
\description{
Update specimens table
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
