% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_manip_db.R
\name{update_plot_data}
\alias{update_plot_data}
\title{Update plot metadata}
\usage{
update_plot_data(
  team_lead = NULL,
  plot_name = NULL,
  country = NULL,
  method = NULL,
  date_y = NULL,
  id_table_plot = NULL,
  new_plot_name = NULL,
  new_team_leader = NULL,
  new_principal_investigator = NULL,
  new_data_manager = NULL,
  new_country = NULL,
  new_ddlat = NULL,
  new_ddlon = NULL,
  new_elevation = NULL,
  new_method = NULL,
  new_province = NULL,
  new_data_provider = NULL,
  new_locality_name = NULL,
  add_backup = TRUE,
  ask_before_update = TRUE
)
}
\arguments{
\item{team_lead}{string name of the team leader of the selected plot}

\item{plot_name}{string plot name of the selected plots}

\item{country}{string country of the selected plots}

\item{method}{string method of the selected plots}

\item{date_y}{integer year of collect of the selected plots}

\item{id_table_plot}{integer id of plot to be updated}

\item{new_plot_name}{string new plot name}

\item{new_team_leader}{string new team leader}

\item{new_principal_investigator}{string}

\item{new_data_manager}{string}

\item{new_country}{string new country}

\item{new_ddlat}{double new latitude in decimal degrees}

\item{new_ddlon}{double new longitude in decimal degrees}

\item{new_elevation}{integer new elevation data}

\item{new_method}{string new method data}

\item{new_province}{string new province data}

\item{add_backup}{logical whether backup of modified data should be recorded}
}
\value{
No return value, new plots are added
}
\description{
Update metadata plot _ only one plot at a time
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
