% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updates_tables_functions.R
\name{update_subplot_data_batch}
\alias{update_subplot_data_batch}
\title{Update subplot data data}
\usage{
update_subplot_data_batch(
  new_data,
  col_names_select = NULL,
  col_names_corresp = NULL,
  id_col = 1,
  launch_update = FALSE,
  add_backup = TRUE
)
}
\arguments{
\item{new_data}{data frame data containing id and new values}

\item{col_names_select}{string plot name of the selected plots}

\item{col_names_corresp}{string of the selected plots}

\item{id_col}{integer indicate which name of col_names_select is the id for matching data}

\item{launch_update}{logical if TRUE updates are performed}

\item{add_backup}{logical whether backup of modified data should be recorded}
}
\value{
No return value individuals updated
}
\description{
Update subplot data plot _ at a time
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
