% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updates_tables_functions.R
\name{update_subplots_table}
\alias{update_subplots_table}
\title{Update subplot_table}
\usage{
update_subplots_table(
  subplots_id = NULL,
  new_id_type_sub_plot = NULL,
  new_typevalue = NULL,
  new_year = NULL,
  new_month = NULL,
  new_day = NULL,
  new_colnam = NULL,
  new_add_people = NULL,
  ask_before_update = TRUE,
  add_backup = TRUE,
  show_results = TRUE
)
}
\arguments{
\item{subplots_id}{integer id of subplot to update}

\item{new_id_type_sub_plot}{id of new type subplot}

\item{new_typevalue}{numeric new value of type subplot}

\item{new_year}{integer new year of subplot}

\item{new_month}{integer new month of subplot}

\item{new_day}{integer new day of subplot}

\item{new_colnam}{character new colnam of subplot}

\item{ask_before_update}{logical TRUE by default, ask for confirmation before updating}

\item{add_backup}{logical TRUE by default, add backup}

\item{show_results}{logical TRUE by default, show the data that has been modified}
}
\value{
No return value individuals updated
}
\description{
Update subplot_table
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
