% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_manip_db.R
\name{update_subplottype_list_table}
\alias{update_subplottype_list_table}
\title{Update subplottype table}
\usage{
update_subplottype_list_table(
  subplottype_searched = NULL,
  id_subplotype = NULL,
  new_subplottype = NULL,
  new_maxallowedvalue = NULL,
  new_minallowedvalue = NULL,
  new_typedescription = NULL,
  new_expectedunit = NULL,
  ask_before_update = TRUE,
  add_backup = TRUE,
  show_results = TRUE
)
}
\arguments{
\item{subplottype_searched}{string genus name searched}

\item{id_subplotype}{string genus name searched}

\item{new_subplottype}{string new subplottype name}

\item{new_maxallowedvalue}{numeric new maxallowedvalue}

\item{new_minallowedvalue}{numeric new minallowedvalue}

\item{new_expectedunit}{string new expectedunit}

\item{ask_before_update}{logical TRUE by default, ask for confirmation before updating}

\item{add_backup}{logical TRUE by default, add backup of modified data}

\item{show_results}{logical TRUE by default, show the data that has been modified}

\item{new_traitdescription}{string new traitdescription}
}
\value{
No return value individuals updated
}
\description{
Update subplottype table
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
