% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_manip_db.R
\name{update_taxa_link_table}
\alias{update_taxa_link_table}
\title{Update id table for taxa}
\usage{
update_taxa_link_table()
}
\description{
Update and rewrite idtaxa to database, require administrator rights
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
