% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_manip_db.R
\name{update_trait_list_table}
\alias{update_trait_list_table}
\title{Update trait_list_table}
\usage{
update_trait_list_table(
  trait_searched = NULL,
  trait_id = NULL,
  new_trait_name = NULL,
  new_relatedterm = NULL,
  new_maxallowedvalue = NULL,
  new_minallowedvalue = NULL,
  new_traitdescription = NULL,
  new_expectedunit = NULL,
  ask_before_update = TRUE,
  add_backup = TRUE,
  show_results = TRUE
)
}
\arguments{
\item{trait_searched}{string genus name searched}

\item{trait_id}{string genus name searched}

\item{new_trait_name}{string new trait name}

\item{new_relatedterm}{string new relatedterm name}

\item{new_maxallowedvalue}{numeric new maxallowedvalue}

\item{new_minallowedvalue}{numeric new minallowedvalue}

\item{new_traitdescription}{string new traitdescription}

\item{new_expectedunit}{string new expectedunit}

\item{ask_before_update}{logical TRUE by default, ask for confirmation before updating}

\item{add_backup}{logical TRUE by default, add backup of modified data}

\item{show_results}{logical TRUE by default, show the data that has been modified}
}
\value{
No return value individuals updated
}
\description{
Update trait_list_table
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
