% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_manip_db.R
\name{update_traits_measures}
\alias{update_traits_measures}
\title{Update traits_measurements table}
\usage{
update_traits_measures(
  new_data,
  trait_values_new_data = NULL,
  col_names_trait_corresp = NULL,
  measures_property_new_data = NULL,
  col_names_property_corresp = NULL,
  id_new_data,
  col_name_id_corresp,
  launch_update = FALSE,
  add_backup = TRUE
)
}
\arguments{
\item{new_data}{string name of the team leader of the selected plot}

\item{trait_values_new_data}{string vector with columns names of new_data that contain traits measures}

\item{col_names_trait_corresp}{string vector with trait names corresponding to trait_values_new_data}

\item{measures_property_new_data}{string vector with columns names new_data others than traits measures values}

\item{col_names_property_corresp}{string vector with corresponding columns names for measures_property_new_data}

\item{id_new_data}{integer which column of new_data contain the trait measure id to match dataset}

\item{col_name_id_corresp}{integer which column of new_data contain the id to match dataset}

\item{launch_update}{logical if TRUE updates are performed}

\item{add_backup}{logical whether backup of modified data should be recorded}
}
\value{
No return value individuals updated
}
\description{
Update traits_measurements table
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
