% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_plots_v2_base.R
\name{validate_query_params}
\alias{validate_query_params}
\title{Valider et normaliser les paramètres de query_plots}
\usage{
validate_query_params(params)
}
\arguments{
\item{params}{Liste de paramètres}
}
\value{
Liste de paramètres validés et normalisés
}
\description{
Fonction utilitaire pour valider les paramètres et résoudre les conflits
}
