% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_manip_db.R
\name{add_individuals}
\alias{add_individuals}
\title{Add new individuals data}
\usage{
add_individuals(new_data, col_names_select, col_names_corresp, id_col)
}
\arguments{
\item{new_data}{string name of the team leader of the selected plot}

\item{col_names_select}{string plot name of the selected plots}

\item{col_names_corresp}{string country of the selected plots}

\item{id_col}{integer indicate which name of col_names_select is the id for matching data}
}
\value{
No return value individuals updated
}
\description{
Add new individuals data
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
