% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_manip_db.R
\name{add_method}
\alias{add_method}
\title{Add a trait in trait list}
\usage{
add_method(new_method = NULL, new_description_method = NULL)
}
\arguments{
\item{new_description_method}{}

\item{new_trait}{string value with new trait descritors - try to avoid space}
}
\value{
nothing
}
\description{
Add trait and associated descriptors in trait list table
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
