% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_manip_db.R
\name{add_method}
\alias{add_method}
\title{Add a method in method list}
\usage{
add_method(new_method = NULL, new_description_method = NULL)
}
\arguments{
\item{new_method}{string value with new method descriptors, avoid space}

\item{new_description_method}{}
}
\value{
nothing
}
\description{
Add method and associated descriptors in method list table
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
