% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_manip_db.R
\name{add_plots}
\alias{add_plots}
\title{Add new plot metadata}
\usage{
add_plots(new_data, col_names_select, col_names_corresp)
}
\arguments{
\item{new_data}{tibble}

\item{col_names_select}{string a vector of string indicating columns names of new_data}

\item{col_names_corresp}{string a vector of string indicating to which columns selected columns of new_data corresponds}
}
\value{
No return value, new plots are added
}
\description{
Add metadata for new plots
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
