% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsplots_features_function.R
\name{add_subplot_features}
\alias{add_subplot_features}
\title{List of subplot features}
\usage{
add_subplot_features(
  new_data,
  col_names_select = NULL,
  col_names_corresp = NULL,
  plot_name_field = NULL,
  id_plot_name = NULL,
  id_plot_name_corresp = "id_table_liste_plots_n",
  subplottype_field,
  features_field = NULL,
  add_data = FALSE,
  ask_before_update = TRUE,
  verbose = TRUE,
  check_existing_data = TRUE
)
}
\value{
A tibble of all subplot features
}
\description{
Provide list of subplot features of plot
}
\author{

}
