% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomic_query_functions.R
\name{add_taxa_table_taxa}
\alias{add_taxa_table_taxa}
\title{Add formatted taxa information}
\usage{
add_taxa_table_taxa(ids = NULL)
}
\arguments{
\item{ids}{vector of idtax_n to retrieve}
}
\value{
tibble
}
\description{
Helper function to add formatted taxonomic names (species, infraspecific, with authors)
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
