% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_manip_db.R
\name{add_trait}
\alias{add_trait}
\title{Add a trait in trait list}
\usage{
add_trait(
  new_trait = NULL,
  new_relatedterm = NULL,
  new_valuetype = NULL,
  new_maxallowedvalue = NULL,
  new_minallowedvalue = NULL,
  new_traitdescription = NULL,
  new_factorlevels = NULL,
  new_expectedunit = NULL,
  new_comments = NULL
)

add_trait(
  new_trait = NULL,
  new_relatedterm = NULL,
  new_valuetype = NULL,
  new_maxallowedvalue = NULL,
  new_minallowedvalue = NULL,
  new_traitdescription = NULL,
  new_factorlevels = NULL,
  new_expectedunit = NULL,
  new_comments = NULL
)
}
\arguments{
\item{new_trait}{string value with new trait descritors - try to avoid space}

\item{new_relatedterm}{string related trait to new trait}

\item{new_valuetype}{string one of following 'numeric', 'integer', 'categorical', 'ordinal', 'logical', 'character'}

\item{new_maxallowedvalue}{numeric if valuetype is numeric, indicate the maximum allowed value}

\item{new_minallowedvalue}{numeric if valuetype is numeric, indicate the minimum allowed value}

\item{new_traitdescription}{string full description of trait}

\item{new_factorlevels}{string a vector of all possible value if valuetype is categorical or ordinal}

\item{new_expectedunit}{string expected unit (unitless if none)}

\item{new_comments}{string any comments}
}
\value{
nothing

nothing
}
\description{
Add trait and associated descriptors in trait list table

See https://terminologies.gfbio.org/terms/ets/pages/index.html for description of each field
}
\details{
Add trait and associated descriptors in trait list table
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}

Gilles Dauby, \email{gilles.dauby@ird.fr}
}
