% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_features_function.R
\name{aggregate_character_features_dt}
\alias{aggregate_character_features_dt}
\title{Aggregate character features using data.table}
\usage{
aggregate_character_features_dt(data, include_census, mode)
}
\description{
Aggregate character features using data.table
}
\keyword{internal}
