% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_features_function.R
\name{aggregate_numeric_features_dt}
\alias{aggregate_numeric_features_dt}
\title{Aggregate numeric features using data.table}
\usage{
aggregate_numeric_features_dt(
  data,
  include_census,
  mode,
  include_issue = FALSE,
  include_measurement_ids = FALSE
)
}
\description{
Aggregate numeric features using data.table
}
\keyword{internal}
