% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections_db.R
\name{call.mydb}
\alias{call.mydb}
\title{Get primary database connection}
\usage{
call.mydb(pass = NULL, user = NULL)
}
\arguments{
\item{pass}{Password (string). If NULL, will prompt using rstudioapi.}

\item{user}{Username (string). If NULL, will prompt using rstudioapi.}

\item{offline}{If TRUE, return offline/mock version of tables.}
}
\value{
A DBI connection or a named list of mock tables (offline mode)
}
\description{
Connects to the 'plots_transects' database. Asks for password and user if not supplied.
}
