% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections_db.R
\name{call.mydb}
\alias{call.mydb}
\title{Get primary database connection (wrapper)}
\usage{
call.mydb(
  pass = NULL,
  user = NULL,
  reset = FALSE,
  retry = TRUE,
  use_env_credentials = FALSE
)
}
\description{
Get primary database connection (wrapper)
}
