% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_manip_db.R
\name{call.mydb.taxa}
\alias{call.mydb.taxa}
\title{Load the taxonomic database}
\usage{
call.mydb.taxa(pass = NULL, user = NULL)
}
\arguments{
\item{pass}{string}

\item{user}{string}
}
\value{
The database is loaded
}
\description{
Load the database and ask for password
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
