% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections_db.R
\name{call.mydb.taxa}
\alias{call.mydb.taxa}
\title{Get taxa database connection}
\usage{
call.mydb.taxa(pass = "Anyuser2022", user = "common", reset = FALSE)
}
\arguments{
\item{pass}{Password (string). If NULL, will prompt using rstudioapi.}

\item{user}{Username (string). If NULL, will prompt using rstudioapi.}
}
\value{
A DBI connection
}
\description{
Connects to the 'rainbio' database.
}
