% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections_db.R
\name{call.mydb.taxa}
\alias{call.mydb.taxa}
\title{Get taxa database connection (wrapper)}
\usage{
call.mydb.taxa(
  pass = NULL,
  user = NULL,
  reset = FALSE,
  retry = TRUE,
  use_env_credentials = FALSE
)
}
\description{
Get taxa database connection (wrapper)
}
