% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections_db.R
\name{connect_database}
\alias{connect_database}
\title{Connect to database}
\usage{
connect_database(
  db_type = c("main", "taxa"),
  pass = NULL,
  user = NULL,
  reset = FALSE,
  retry = TRUE,
  use_env_credentials = FALSE
)
}
\arguments{
\item{db_type}{One of `"main"` or `"taxa"`.}

\item{pass}{Password. If NULL, will check environment then prompt.}

\item{user}{Username. If NULL, will check environment then prompt.}

\item{reset}{If TRUE, forces new credential prompt.}

\item{retry}{If TRUE, retry on failure.}

\item{use_env_credentials}{If TRUE, tries to use MYDB_USER and MYDB_PASS from .Renviron (default: FALSE)}
}
\value{
A database connection object.
}
\description{
Generic function to connect to main or taxa database
}
