% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_manip_db.R
\name{divid_plot}
\alias{divid_plot}
\title{Divide 1 ha square plots into 25 squares subplots of 400m² following a regular 5*5 grid}
\usage{
divid_plot(
  coordinates_sf,
  plot_name = "plot_name",
  crs = NULL,
  long = "typevalue_ddlon",
  lat = "typevalue_ddlat",
  XRel = "Xrel",
  YRel = "Yrel",
  order = 1
)
}
\arguments{
\item{plot_name}{the name as character of the column where the plot name are stored. Default value is 'plot_name'.}

\item{crs}{the crs. ex : 'EPSG:32633'}

\item{long}{the name as character of the column where the longitude are stored. Default value is 'typevalue_ddlon'}

\item{lat}{the name as character of the column where the longitude are stored. Default value is "typevalue_ddlat"}

\item{XRel}{the name as character of the column where the longitude are stored. Default value is 'XRel'}

\item{YRel}{the name as character of the column where the longitude are stored. Default value is 'YRel'}

\item{order}{1 or 2. Condition the subplot name order. If the line 0 in Y goes a bit to the south choose 1, and 2 if it goes a bit to the North.}

\item{coordinates}{a dataframe with long and lat coord for all the 20 'jalons' around the plot.}
}
\value{
A \code{sf} object with the 25 subplots geometries with 2 fields : sous_plot_name and plot_name for each plot.
}
\description{
Divide 1 ha square plots into 25 squares subplots of 400m² following a regular 5*5 grid
}
\details{
The function works well with the $coordinates dataframe export by the query_plots function. The subplots names
could be in the wrong order depending on the orientation of the plot. To fit this problem, change the value of 'order'
}
\examples{
library(mapview)
myplots <- query_plots(locality_name = "Mbalmayo", show_all_coordinates = TRUE, map = T, extract_individuals = F, extract_traits = F, method = '1ha-IRD')
myplots_coord <- myplots$coordinates
coordinates_sf <- myplots_coord

test1 <- divid_plot(coordinates_sf = coordinates_sf,
plot_name = 'id_table_liste_plots',
crs = 'EPSG32633',
long = "typevalue_ddlon",
lat = "typevalue_ddlat",
XRel = "Xrel",
YRel = "Yrel",
order = 1))

# To modify the order of the subplot names

test2 <- divid_plot(coordinates_sf = coordinates_sf,
plot_name = 'id_table_liste_plots',
crs = 'EPSG32633',
long = "typevalue_ddlon",
lat = "typevalue_ddlat",
XRel = "Xrel",
YRel = "Yrel",
order = 2))

mapview::mapview(test1, zcol = 'sous_plot_name')
mapview::mapview(test2, zcol = 'sous_plot_name')

}
