% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.add_modif_field}
\alias{.add_modif_field}
\title{Internal function}
\usage{
.add_modif_field(dataset)
}
\arguments{
\item{dataset}{string tibble to add dates fields}
}
\value{
vector
}
\description{
Add modification day month and year column before adding/updating
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
