% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R, R/updates_tables_functions.R
\name{.comp_print_vec}
\alias{.comp_print_vec}
\title{Compare two row-tibbles and generate HTML with differences}
\usage{
.comp_print_vec(vec_1, vec_2)

.comp_print_vec(vec_1, vec_2)
}
\arguments{
\item{vec_1}{A tibble with one row}

\item{vec_2}{A tibble with one row}
}
\value{
A list: (1) tibble of differing columns, (2) HTML table highlighting differences

A list: (1) tibble of differing columns, (2) HTML table highlighting differences
}
\description{
Compare two row-tibbles and generate HTML with differences

Compare two row-tibbles and generate HTML with differences
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
