% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_manip_db.R
\name{.delete_link_individual_specimen}
\alias{.delete_link_individual_specimen}
\title{Delete an entry in trait measurement table}
\usage{
.delete_link_individual_specimen(id_ind = NULL, id_specimen = NULL)
}
\arguments{
\item{id_ind}{integer}

\item{id_specimen}{integer}

\item{id}{integer}
}
\value{
No values
Delete an entry in trait measurement table

Delete an entry in trait measurement table using id for selection

No values
}
\description{
Delete an entry in diconame table using id for selection
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}

Gilles Dauby, \email{gilles.dauby@ird.fr}
}
