% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_table_functions.R
\name{.find_cat}
\alias{.find_cat}
\title{Internal function}
\usage{
.find_cat(
  value_to_search,
  compared_table,
  column_name,
  prompt_message = "Choose"
)
}
\arguments{
\item{value_to_search}{string vector of one element}

\item{compared_table}{tibble with one column where the value should be compared}

\item{column_name}{string name of the column of compared_table}
}
\value{
vector
}
\description{
Semi automatic matching with a table for comparison
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
