% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_table_functions.R
\name{.link_table}
\alias{.link_table}
\title{Internal function}
\usage{
.link_table(
  data_stand,
  column_searched,
  column_name,
  id_field,
  id_table_name,
  db_connection = mydb,
  table_name
)
}
\arguments{
\item{data_stand}{tibble}

\item{column_searched}{string vector}

\item{id_field}{name of the column that will store the id}
}
\value{
vector
}
\description{
Compute
}
\examples{
# .link_table(data_stand = data_stand, column_searched = "method", column_name = "method", id_field = "id_method", db_connection = mydb, table_name = "methodslist")


}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
