% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_table_functions.R
\name{.link_table}
\alias{.link_table}
\title{Internal function}
\usage{
.link_table(
  data_stand,
  column_searched,
  column_name,
  id_field,
  id_table_name,
  db_connection = mydb,
  table_name,
  keep_columns = NULL,
  keep_original_value = FALSE
)
}
\arguments{
\item{data_stand}{tibble}

\item{column_searched}{string vector}

\item{column_name}{name of the column that will store the id}

\item{id_field}{string name of the column of the output that will contain the id}

\item{id_table_name}{string name of the database table that contain the id}

\item{db_connection}{PqConnection  connection of the database}

\item{table_name}{string name of the table in the database to search into}

\item{keep_columns}{string vector of the columns in the database table to keep in the output, by default NULL}
}
\value{
vector
}
\description{
Compute
}
\examples{
# .link_table(data_stand = data_stand, column_searched = "method", column_name = "method", id_field = "id_method", db_connection = mydb, table_name = "methodslist")


}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
