% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_table_functions.R
\name{.link_trait}
\alias{.link_trait}
\title{Internal function}
\usage{
.link_trait(data_stand, trait, column_name = "trait", issues = NULL)
}
\arguments{
\item{data_stand}{tibble}

\item{trait}{string vector}

\item{issues}{string vector}
}
\value{
vector
}
\description{
Compute
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
