% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_manip_db.R
\name{launch_query_tax_app}
\alias{launch_query_tax_app}
\title{Launch shiny app for taxonomic standardization}
\usage{
launch_query_tax_app()
}
\value{
Launch in web browser the app
}
\description{
Tool for standardizing taxonomy according to the taxonomic backbone (step needed before importing data to databases)
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
