% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomic_update_functions.R
\name{merge_individuals_taxa}
\alias{merge_individuals_taxa}
\title{Merge individual records with taxonomic information}
\usage{
merge_individuals_taxa(
  id_individual = NULL,
  id_plot = NULL,
  id_tax = NULL,
  clean_columns = TRUE
)
}
\arguments{
\item{id_individual}{vector of individual IDs to filter}

\item{id_plot}{vector of plot IDs to filter}

\item{id_tax}{vector of taxa IDs to filter final results}

\item{clean_columns}{logical, whether to remove redundant columns}
}
\value{
A tibble with individuals and resolved taxonomy
}
\description{
Merges individual tree records with resolved taxonomy from both direct links
and specimen links, handling synonym resolution
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
