% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_manip_db.R
\name{merge_individuals_taxa}
\alias{merge_individuals_taxa}
\title{Merge individuals and taxonomic backbone}
\usage{
merge_individuals_taxa(id_individual = NULL, id_plot = NULL, id_tax = NULL)
}
\arguments{
\item{id_individual}{numeric id of individuals to be extracted if any}

\item{id_plot}{numeric id of plots to be extracted if any}
}
\value{
A src object with merged taxa
}
\description{
Merge idividuals and taxonomic backbone
#'
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
