% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{query_fuzzy_match}
\alias{query_fuzzy_match}
\title{Query fuzzy match}
\usage{
query_fuzzy_match(tbl, field, values_q, con)
}
\arguments{
\item{tbl}{tibble with one field listing names to be searched}

\item{field}{string column name to be search}

\item{values_q}{string names to be searched}

\item{con}{PqConnection connection to RPostgres database}
}
\value{
A list of two elements, one with the extract if any, two with the names with id not NA when matched
}
\description{
Extract from a sql database an exact match on a given field
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
