% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_features_function.R
\name{query_individual_features}
\alias{query_individual_features}
\title{Query for features linked to individual}
\usage{
query_individual_features(
  id = NULL,
  multiple_census = FALSE,
  id_traits = NULL,
  pivot_table = TRUE,
  extract_trait_measures_features = FALSE,
  extract_linked_individuals = FALSE
)
}
\arguments{
\item{id}{numeric vector of id from individuals}

\item{multiple_census}{boolean whether multiple census should be shown}

\item{id_traits}{numeric vector of id from traits/features}

\item{pivot_table}{boolean whether results should be pivoted into multiple columns}

\item{extract_trait_measures_features}{boolean whether adding additional features linked to traits/features, only if pivot_table is FALSE}
}
\value{
tibble
}
\description{
Query for features linked to a given individual in inventories
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
