% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_manip_db.R
\name{query_plots}
\alias{query_plots}
\title{List and map selected plots}
\usage{
query_plots(team_lead = NULL, plot_name = NULL, country = NULL,
  province = NULL, method = NULL, date_y = NULL,
  extract_individuals = FALSE, map = FALSE, zoom = 0, label = T,
  try_optimal_label = F, map_type = "mapview")
}
\arguments{
\item{team_lead}{string}

\item{plot_name}{string}

\item{country}{string}

\item{province}{string}

\item{method}{string}

\item{date_y}{integer}

\item{extract_individuals}{logical}

\item{map}{logical}

\item{zoom}{integer positive values indicating zoom level for ggplot style map}

\item{label}{logical for ggplot map, whether adding label or not}

\item{try_optimal_label}{logical}

\item{map_type}{string mapview or ggplot}
}
\value{
A tibble of all method

A tibble of plots or individuals if extract_individuals is TRUE
}
\description{
Provide and map list of selected plots
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
