% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_manip_db.R
\name{query_plots}
\alias{query_plots}
\title{List, extract map selected plots}
\usage{
query_plots(
  team_lead = NULL,
  plot_name = NULL,
  tag = NULL,
  country = NULL,
  province = NULL,
  locality_name = NULL,
  method = NULL,
  date_y = NULL,
  extract_individuals = FALSE,
  map = FALSE,
  zoom = 0,
  label = T,
  try_optimal_label = F,
  map_type = "mapview",
  id_individual = NULL,
  id_plot = NULL,
  id_tax = NULL,
  show_multiple_census = FALSE,
  show_all_coordinates = FALSE,
  remove_ids = TRUE,
  collapse_multiple_val = FALSE,
  extract_traits = TRUE,
  traits_to_genera = FALSE,
  wd_fam_level = FALSE,
  include_liana = FALSE,
  extract_subplot_features = TRUE
)
}
\arguments{
\item{team_lead}{string fuzzy person name to look for}

\item{plot_name}{string fuzzy plot name to look for}

\item{tag}{numeric exact tag number of the plot}

\item{country}{string fuzzy country name to look for}

\item{province}{string fuzzy province name to look for}

\item{locality_name}{string fuzzy locality_name name to look for}

\item{method}{stringfuzzy method name to look for}

\item{date_y}{integer year of collect}

\item{extract_individuals}{logical whether individuals instead of plot list as output}

\item{map}{logical whether map whould be produced}

\item{zoom}{integer positive values indicating zoom level for ggplot style map}

\item{label}{logical for ggplot map, whether adding label or not}

\item{try_optimal_label}{logical if map is ggplot}

\item{map_type}{string mapview or ggplot, mapview by default}

\item{id_individual}{numeric id of individual to be extracted}

\item{id_plot}{numeric id of plot to be extracted}

\item{show_multiple_census}{logical whether multiple census should be shown, by default FALSE}

\item{remove_ids}{logical remove all ids columns, by default TRUE}

\item{collapse_multiple_val}{logical whether multiple traits measures should be collapsed (resulting values as character, separated by dash)}

\item{extract_traits}{whether species level traits should be extracted as well}

\item{traits_to_genera}{if species-level traits should be extrapolated to genus level, by default is FALSE}

\item{wd_fam_level}{logical, if wood density should be given at family level or not NOT YET FULLY AVAILABLE}

\item{include_liana}{logical, if liana should be included, by default FALSE}
}
\value{
A tibble of all plots or individuals

A tibble of plots or individuals if extract_individuals is TRUE
}
\description{
Provide and map list of selected plots, including associated traits for individuals
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
