% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsplots_features_function.R
\name{query_subplots}
\alias{query_subplots}
\title{List, selected subplots}
\usage{
query_subplots(
  ids_plots = NULL,
  ids_subplots = NULL,
  plot_name = NULL,
  country = NULL,
  locality_name = NULL,
  method = NULL,
  subtype = NULL,
  verbose = TRUE,
  extract_subplots_obs_features = FALSE
)
}
\arguments{
\item{ids_subplots}{integer}

\item{plot_name}{string fuzzy plot name to look for}

\item{country}{string fuzzy country name to look for}

\item{locality_name}{string fuzzy locality_name name to look for}

\item{method}{stringfuzzy method name to look for}

\item{subtype}{string subtype to select}

\item{verbose}{logical}

\item{extract_subplots_obs_features}{logical}

\item{id_plots}{integer}

\item{tag}{numeric exact tag number of the plot}
}
\value{
A tibble of all subplots
}
\description{
Table of subplot for selected plots
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
