% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_manip_db.R
\name{query_subplots}
\alias{query_subplots}
\title{List, selected subplots}
\usage{
query_subplots(
  ids_plots = NULL,
  team_lead = NULL,
  plot_name = NULL,
  tag = NULL,
  country = NULL,
  province = NULL,
  locality_name = NULL,
  method = NULL,
  subtype = NULL,
  verbose = TRUE
)
}
\arguments{
\item{team_lead}{string fuzzy person name to look for}

\item{plot_name}{string fuzzy plot name to look for}

\item{tag}{numeric exact tag number of the plot}

\item{country}{string fuzzy country name to look for}

\item{province}{string fuzzy province name to look for}

\item{locality_name}{string fuzzy locality_name name to look for}

\item{method}{stringfuzzy method name to look for}

\item{subtype}{string subtype to select}

\item{verbose}{logical}

\item{id_plots}{integer}
}
\value{
A tibble of all subplots
}
\description{
Table of subplot for selected plots
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
