% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_manip_db.R
\name{query_taxa}
\alias{query_taxa}
\title{List, extract taxa}
\usage{
query_taxa(
  class = NULL,
  family = NULL,
  genus = NULL,
  order = NULL,
  species = NULL,
  tax_nam01 = NULL,
  tax_nam02 = NULL,
  only_genus = FALSE,
  only_family = FALSE,
  only_class = FALSE,
  habit = NULL,
  ids = NULL,
  verbose = TRUE,
  exact_match = FALSE,
  check_synonymy = TRUE,
  extract_traits = FALSE
)
}
\arguments{
\item{team_lead}{string fuzzy person name to look for}

\item{plot_name}{string fuzzy plot name to look for}

\item{tag}{numeric exact tag number of the plot}

\item{country}{string fuzzy country name to look for}

\item{province}{string fuzzy province name to look for}

\item{locality_name}{string fuzzy locality_name name to look for}

\item{method}{stringfuzzy method name to look for}

\item{date_y}{integer year of collect}

\item{extract_individuals}{logical whether individuals instead of plot list as output}

\item{map}{logical whether map whould be produced}

\item{zoom}{integer positive values indicating zoom level for ggplot style map}

\item{label}{logical for ggplot map, whether adding label or not}

\item{try_optimal_label}{logical if map is ggplot}

\item{map_type}{string mapview or ggplot, mapview by default}

\item{id_individual}{numeric id of individual to be extracted}

\item{id_plot}{numeric id of plot to be extracted}

\item{show_multiple_census}{logical whether multiple census should be shown, by default FALSE}

\item{remove_ids}{logical remove all ids columns, by default TRUE}

\item{collapse_multiple_val}{logical whether multiple traits measures should be collapsed (resulting values as character, separated by dash)}
}
\value{
A tibble of all taxa

A tibble of plots or individuals if extract_individuals is TRUE
}
\description{
Provide list of selected taxa
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
