% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_manip_db.R
\name{query_taxa}
\alias{query_taxa}
\title{List, extract taxa}
\usage{
query_taxa(
  class = NULL,
  family = NULL,
  genus = NULL,
  order = NULL,
  species = NULL,
  only_genus = FALSE,
  only_family = FALSE,
  only_class = FALSE,
  ids = NULL,
  verbose = TRUE,
  exact_match = FALSE,
  check_synonymy = TRUE,
  extract_traits = TRUE
)
}
\arguments{
\item{class}{character string of class}

\item{family}{string}

\item{genus}{string}

\item{order}{string}

\item{species}{string genus followed by species name separated by one space}

\item{only_genus}{logical}

\item{only_family}{logical}

\item{only_class}{logical}

\item{ids}{integer id of searched taxa}

\item{verbose}{logical}

\item{exact_match}{logical}

\item{check_synonymy}{logical}

\item{extract_traits}{logical}

\item{tax_nam01}{string}

\item{tax_nam02}{string}
}
\value{
A tibble of all taxa

A tibble of plots or individuals if extract_individuals is TRUE
}
\description{
Provide list of selected taxa
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
