% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_traits_function.R
\name{query_traits_measures}
\alias{query_traits_measures}
\title{Query for taxa level traits}
\usage{
query_traits_measures(
  idtax = NULL,
  idtax_good = NULL,
  add_taxa_info = FALSE,
  id_trait = NULL,
  trait_cat_mode = "most_frequent",
  verbose = TRUE,
  pivot_table = TRUE,
  include_remarks = FALSE,
  extract_trait_measures_features = FALSE
)
}
\arguments{
\item{idtax}{vector of idtax_n}

\item{idtax_good}{vector of idtax_good; NULL if no synonym}

\item{add_taxa_info}{logical}

\item{id_trait}{integer indicating the id of the trait to look for}

\item{trait_cat_mode}{vector string if "most_frequent" then the most frequent value for categorical trait is given, if "all_unique" then all unique value separated by comma}

\item{verbose}{logical}

\item{pivot_table}{logical}

\item{include_remarks}{logical}

\item{extract_trait_measures_features}{logical}
}
\value{
vector
}
\description{
Query for traits for given taxa, include synonyms or not
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
