% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_manip_db.R
\name{query_traits_measures}
\alias{query_traits_measures}
\title{Query for traits}
\usage{
query_traits_measures(
  idtax,
  idtax_good = NULL,
  add_taxa_info = FALSE,
  trait_cat_mode = "most_frequent",
  verbose = TRUE
)
}
\arguments{
\item{idtax}{vector of idtax_n}

\item{idtax_good}{vector of idtax_good; NA if no synonym}

\item{add_taxa_info}{logical}

\item{trait_cat_mode}{vector string if "most_frequent" then the most frequent value for categorical trait is given, if "all_unique" then all unique value separated by comma}

\item{verbose}{logical}
}
\value{
vector
}
\description{
Query for traits for given taxa, include synonyms or not
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
