% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_features_function.R
\name{query_traits_measures_features}
\alias{query_traits_measures_features}
\title{List, selected trait measures features}
\usage{
query_traits_measures_features(id_trait_measures = NULL)
}
\arguments{
\item{id_trait_measures}{integer}
}
\value{
A tibble of all subplots
}
\description{
Table of srait measures features
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
