% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_manip_db.R
\name{replace_NA}
\alias{replace_NA}
\title{Internal function}
\usage{
replace_NA(vec, inv = FALSE)
}
\arguments{
\item{vec}{vector}
}
\value{
tibble
}
\description{
Replace NA by 9999
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
