% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_manip_db.R
\name{traits_list}
\alias{traits_list}
\title{List of trait and features potentially liked to individual}
\usage{
traits_list(id_trait = NULL)
}
\value{
A tibble of all traits and features that can be linked to individual
}
\description{
Provide list of traits and features available
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
