% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_manip_db.R
\name{try_open_postgres_table}
\alias{try_open_postgres_table}
\title{Open postgresql database table}
\usage{
try_open_postgres_table(table, con)
}
\arguments{
\item{table}{string}

\item{con}{connection to database}
}
\description{
Access to postgresql database table with repeating if no successful
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
