% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomic_update_functions.R
\name{update_taxa_link_table}
\alias{update_taxa_link_table}
\title{Update local taxonomy link table}
\usage{
update_taxa_link_table()
}
\value{
NULL (updates database table)
}
\description{
Updates the table_idtax in main database with latest synonym information
from the taxa database
}
\author{
Gilles Dauby, \email{gilles.dauby@ird.fr}
}
